# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for Telegram WebApp (TAM).",
        "title": "TAM Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/channels/add-channel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new Telegram channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Add channel",
                "parameters": [
                    {
                        "description": "Channel details",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.AddChannelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.CreateChannelResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/channels/list-channels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get listed channels with pagination and optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "List channels",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (max 100)",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "DESC",
                        "description": "Sort order (ASC or DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter only user's channels (admin/owner)",
                        "name": "myChannels",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.ListChannelsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/channels/provide-channel-data": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update channel details (description, tags, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Provide channel data",
                "parameters": [
                    {
                        "description": "Channel data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.ProvideChannelDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.ChannelResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Channel not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/channels/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get channel details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Get channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.GetChannelResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Channel not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete channel by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Delete channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_wickedv43_TAM-backend_internal_common.AddChannelRequest": {
            "type": "object",
            "properties": {
                "channel_username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.ChannelPrices": {
            "type": "object",
            "properties": {
                "post_1_24": {
                    "type": "number"
                },
                "post_2_48": {
                    "type": "number"
                },
                "post_3_72": {
                    "type": "number"
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.CreateChannelResponse": {
            "type": "object",
            "properties": {
                "avg_post_views": {
                    "type": "integer"
                },
                "commentary": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "first_post_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_listed": {
                    "type": "boolean"
                },
                "main_language": {
                    "type": "string"
                },
                "median_post_views": {
                    "type": "integer"
                },
                "notifications_on": {
                    "type": "number"
                },
                "premium_subscribers": {
                    "type": "integer"
                },
                "prices": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.ChannelPrices"
                },
                "restricted_from": {
                    "type": "string"
                },
                "restricted_till": {
                    "type": "string"
                },
                "restriction_reason": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_database_ent_channel.RestrictionReason"
                },
                "stats_updated_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common_constants_channel_status.ChannelStatus"
                },
                "subscribers": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tg_description": {
                    "type": "string"
                },
                "tg_id": {
                    "type": "integer"
                },
                "tg_name": {
                    "type": "string"
                },
                "tg_picture": {
                    "type": "string"
                },
                "tg_username": {
                    "type": "string"
                },
                "total_posts": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.GetChannelResponse": {
            "type": "object",
            "properties": {
                "canEdit": {
                    "type": "boolean"
                },
                "channel": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.ChannelResponse"
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.ListChannelsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.ChannelResponseWithoutStats"
                    }
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.ProvideChannelDataRequest": {
            "type": "object",
            "properties": {
                "channel_description": {
                    "type": "string"
                },
                "channel_id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "prices": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.ChannelPrices"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_database_ent_channel.RestrictionReason": {
            "type": "string",
            "enum": [
                "unspecified",
                "unspecified",
                "cant_verify"
            ],
            "x-enum-varnames": [
                "DefaultRestrictionReason",
                "RestrictionReasonUnspecified",
                "RestrictionReasonCantVerify"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}